requirements.txt
flask
psycopg2-binary
pandas
openpyxl
anthropic
python-dotenv
